add simple tests and rename module to all lower
