+++ tests/sample_data/brt_simple.gpx
@@ -1,0 +1,4 @@
+<gpx xmlns="http://www.topografix.com/GPX/1/1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.topografix.com/GPX/1/1 http://www.topografix.com/GPX/1/1/gpx.xsd" version="1.1" creator="maps.ie"><metadata><name>simple_brt</name><desc></desc><time>2021-06-09 07:34:09</time></metadata><trk><name></name><desc>stroke=#2593d9
+fill=#2593d9
+stroke-opacity=1
+stroke-width=3</desc><trkseg><trkpt lat="-22.956271" lon="-43.198156"><ele>-1</ele></trkpt><trkpt lat="-22.956306" lon="-43.198142"><ele>23.53</ele></trkpt><trkpt lat="-22.955748" lon="-43.196574"><ele>20.54</ele></trkpt><trkpt lat="-22.955627" lon="-43.196201"><ele>20.73</ele></trkpt><trkpt lat="-22.954835" lon="-43.194033"><ele>15.33</ele></trkpt><trkpt lat="-22.953085" lon="-43.189064"><ele>22.68</ele></trkpt><trkpt lat="-22.953085" lon="-43.189064"><ele>22.68</ele></trkpt><trkpt lat="-22.952539" lon="-43.187477"><ele>19.47</ele></trkpt><trkpt lat="-22.951471" lon="-43.184561"><ele>15.94</ele></trkpt></trkseg></trk></gpx>
