+++ tests/sample_data/sample_gtfs/agency.txt
@@ -1,0 +1,2 @@
+agency_id,agency_name,agency_url,agency_timezone,agency_lang,agency_phone,agency_fare_url
+148,BRT Rio,http://brt.rio/,America/Sao_Paulo,PT,0800 886 1000,https://www.cartaoriocard.com.br/rcc/institucional/tarifas
