+++ tests/sample_data/sample_gtfs/calendar.txt
@@ -1,0 +1,4 @@
+service_id,monday,tuesday,wednesday,thursday,friday,saturday,sunday,start_date,end_date
+S,0,0,0,0,0,1,0,20210101,20210102
+D,0,0,0,0,0,0,1,20210101,20210102
+U,1,1,1,1,1,0,0,20210101,20210102
