+++ tests/sample_data/sample_gtfs/frequencies.txt
@@ -1,0 +1,3 @@
+old_trip_id,trip_id,start_time,end_time,headway_secs,calendar
+000,1,03:00:00,04:00:00,3600,U
+000,2,04:00:00,05:00:00,3600,U
