+++ tests/sample_data/sample_gtfs/routes.txt
@@ -1,0 +1,2 @@
+old_route_id,route_id,agency_id,route_short_name,route_long_name,route_type,route_color,route_text_color
+000,01,000,BRT_SIMPLE,brt_simple_route,3,000080,FFFFFF
