+++ tests/sample_data/sample_gtfs/shapes.txt
@@ -1,0 +1,7 @@
+shape_id,shape_pt_lat,shape_pt_lon,shape_pt_sequence
+001,-22.9563446, -43.1982127,0
+001,-22.956344300312733, -43.20796352014157,1
+001,-22.956344225390918, -43.212838930206985,2
+002,-22.9563446, -43.1982127,2
+002,-22.956344300312733, -43.20796352014157,1
+002,-22.956344225390918, -43.212838930206985,0
