+++ tests/sample_data/sample_gtfs/simple_realized_trips.csv
@@ -1,0 +1,5 @@
+vehicle_id,departure_datetime,arrival_datetime,departure_id,arrival_id,distance,elapsed_time,average_speed,trajectory_tipe,trip_id,trip_short_name
+none,2021-01-01 03:00:00,2021-01-01 03:30:00,0,2,1500.0,0.5,3.0,complete_trip,1,
+none,2021-01-01 03:00:00,2021-01-01 03:30:00,0,2,1500.0,0.5,3.0,complete_trip,1,
+none,2021-01-01 03:00:00,2021-01-01 03:30:00,2,0,1500.0,0.5,3.0,complete_trip,2,
+none,2021-01-01 03:00:00,2021-01-01 03:30:00,2,0,1500.0,0.5,3.0,complete_trip,2,
