+++ tests/sample_data/sample_gtfs/stop_times.txt
@@ -1,0 +1,7 @@
+trip_id,arrival_time,departure_time,stop_id,stop_sequence
+1,01:00:00,01:00:00,0,0
+1,01:20:00,01:20:00,1,1
+1,01:30:00,01:30:00,2,2
+2,01:00:00,01:00:00,2,0
+2,01:20:00,01:20:00,1,1
+2,01:30:00,01:30:00,0,2
